moved RuleSet definition to its own file
